refactor: Rule의 adjustSumByAce() 메서드를 Hands로 이동
