+++ src/test/java/blackjack/domain/gambler/HandsTest.java
@@ -93,4 +93,12 @@ class HandsTest {
         return Stream.of(arguments1, arguments2, arguments3, arguments4, arguments5, arguments6, arguments7,
                 arguments8);
     }
+
+    @DisplayName("에이스를 고려해서 총합을 조정한다")
+    @CsvSource(value = {"18,0,18", "23,1,13", "21,1,21", "18,1,18", "24,2,14", "22,2,12"})
+    @ParameterizedTest
+    void adjustSumByAceTest(int sum, int aceCount, int expected) {
+        Hands hands = new Hands();
+        assertThat(hands.adjustSumByAce(sum, aceCount)).isEqualTo(expected);
+    }
 }
